[INLONG-5030][Manager] Use SortConfigListener replace the LightGroupSortListener
