+++ inlong-manager/manager-common/src/main/java/org/apache/inlong/manager/common/pojo/source/tubemq/TubeMQSource.java
@@ -58,7 +58,7 @@ public class TubeMQSource extends StreamSource {
     private String sessionKey;
 
     /**
-     * The tubemq consumers use this tid set to filter records reading from server.
+     * The TubeMQ consumers use this tid set to filter records reading from server.
      */
     @ApiModelProperty("Tid of the TubeMQ")
     private TreeSet<String> tid;
