+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sink/StreamSinkService.java
@@ -18,13 +18,16 @@
 package org.apache.inlong.manager.service.sink;
 
 import com.github.pagehelper.PageInfo;
+import org.apache.inlong.manager.common.pojo.group.InlongGroupInfo;
 import org.apache.inlong.manager.common.pojo.sink.SinkApproveDTO;
 import org.apache.inlong.manager.common.pojo.sink.SinkBriefResponse;
 import org.apache.inlong.manager.common.pojo.sink.SinkPageRequest;
 import org.apache.inlong.manager.common.pojo.sink.SinkRequest;
 import org.apache.inlong.manager.common.pojo.sink.StreamSink;
+import org.apache.inlong.manager.common.pojo.stream.InlongStreamInfo;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * Service layer interface for stream sink
@@ -65,6 +68,15 @@ public interface StreamSinkService {
      * @return Sink info list.
      */
     List<SinkBriefResponse> listBrief(String groupId, String streamId);
+
+    /**
+     * Get the StreamSink Map by the inlong group info and inlong stream info list.
+     *
+     * @param groupInfo inlong group info
+     * @param streamInfos inlong stream info list
+     * @return map of StreamSink list, key-inlongStreamId, value-StreamSinkList
+     */
+    Map<String, List<StreamSink>> getSinksMap(InlongGroupInfo groupInfo, List<InlongStreamInfo> streamInfos);
 
     /**
      * Query the number of undeleted sink info based on inlong group and inlong stream id
