--- inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sort/light/LightGroupSortSelector.java
@@ -1,51 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.inlong.manager.service.sort.light;
-
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.collections.CollectionUtils;
-import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
-import org.apache.inlong.manager.common.pojo.stream.InlongStreamInfo;
-import org.apache.inlong.manager.common.pojo.workflow.form.process.LightGroupResourceProcessForm;
-import org.apache.inlong.manager.common.pojo.workflow.form.process.ProcessForm;
-import org.apache.inlong.manager.workflow.WorkflowContext;
-import org.apache.inlong.manager.workflow.event.EventSelector;
-
-import java.util.List;
-
-/**
- * Selector of light group sort.
- */
-@Slf4j
-public class LightGroupSortSelector implements EventSelector {
-
-    @Override
-    public boolean accept(WorkflowContext context) {
-        ProcessForm processForm = context.getProcessForm();
-        if (!(processForm instanceof LightGroupResourceProcessForm)) {
-            return false;
-        }
-        LightGroupResourceProcessForm lightGroupResourceProcessForm = (LightGroupResourceProcessForm) processForm;
-        List<InlongStreamInfo> streamInfos = lightGroupResourceProcessForm.getStreamInfos();
-        if (CollectionUtils.isEmpty(streamInfos)) {
-            log.warn(ErrorCodeEnum.STREAM_NOT_FOUND.getMessage());
-            return false;
-        }
-        return true;
-    }
-}
