+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/source/StreamSourceService.java
@@ -18,11 +18,14 @@
 package org.apache.inlong.manager.service.source;
 
 import com.github.pagehelper.PageInfo;
+import org.apache.inlong.manager.common.pojo.group.InlongGroupInfo;
 import org.apache.inlong.manager.common.pojo.source.SourcePageRequest;
 import org.apache.inlong.manager.common.pojo.source.SourceRequest;
 import org.apache.inlong.manager.common.pojo.source.StreamSource;
+import org.apache.inlong.manager.common.pojo.stream.InlongStreamInfo;
 
 import java.util.List;
+import java.util.Map;
 
 /**
  * Service layer interface for stream source
@@ -54,6 +57,18 @@ public interface StreamSourceService {
      * @return Source info list.
      */
     List<StreamSource> listSource(String groupId, String streamId);
+
+    /**
+     * Get the StreamSource Map by the inlong group info and inlong stream info list.
+     * <p/>
+     * If the group mode is LIGHTWEIGHT, means not using any MQ as a cached source, then just get all related sources.
+     * Otherwise, if the group mode is NORMAL, need get the cached MQ sources.
+     *
+     * @param groupInfo inlong group info
+     * @param streamInfos inlong stream info list
+     * @return map of StreamSource list, key-inlongStreamId, value-StreamSourceList
+     */
+    Map<String, List<StreamSource>> getSourcesMap(InlongGroupInfo groupInfo, List<InlongStreamInfo> streamInfos);
 
     /**
      * Query the number of undeleted source info based on inlong group and inlong stream id.
