+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/source/StreamSourceServiceImpl.java
@@ -24,14 +24,17 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.inlong.manager.common.consts.InlongConstants;
 import org.apache.inlong.manager.common.enums.ErrorCodeEnum;
 import org.apache.inlong.manager.common.enums.GroupStatus;
 import org.apache.inlong.manager.common.enums.SourceStatus;
 import org.apache.inlong.manager.common.enums.SourceType;
 import org.apache.inlong.manager.common.exceptions.BusinessException;
+import org.apache.inlong.manager.common.pojo.group.InlongGroupInfo;
 import org.apache.inlong.manager.common.pojo.source.SourcePageRequest;
 import org.apache.inlong.manager.common.pojo.source.SourceRequest;
 import org.apache.inlong.manager.common.pojo.source.StreamSource;
+import org.apache.inlong.manager.common.pojo.stream.InlongStreamInfo;
 import org.apache.inlong.manager.common.pojo.stream.StreamField;
 import org.apache.inlong.manager.common.util.CommonBeanUtils;
 import org.apache.inlong.manager.common.util.Preconditions;
@@ -51,8 +54,10 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
  * Implementation of source service interface
@@ -63,7 +68,7 @@ public class StreamSourceServiceImpl implements StreamSourceService {
     private static final Logger LOGGER = LoggerFactory.getLogger(StreamSourceServiceImpl.class);
 
     @Autowired
-    private SourceOperatorFactory operationFactory;
+    private SourceOperatorFactory operatorFactory;
     @Autowired
     private StreamSourceEntityMapper sourceMapper;
     @Autowired
@@ -91,13 +96,13 @@ public class StreamSourceServiceImpl implements StreamSourceService {
 
         // According to the source type, save source information
         String sourceType = request.getSourceType();
-        StreamSourceOperator operation = operationFactory.getInstance(SourceType.forType(sourceType));
+        StreamSourceOperator sourceOperator = operatorFactory.getInstance(SourceType.forType(sourceType));
         // Remove id in sourceField when save
         List<StreamField> streamFields = request.getFieldList();
         if (CollectionUtils.isNotEmpty(streamFields)) {
             streamFields.forEach(streamField -> streamField.setId(null));
         }
-        int id = operation.saveOpt(request, groupEntity.getStatus(), operator);
+        int id = sourceOperator.saveOpt(request, groupEntity.getStatus(), operator);
 
         LOGGER.info("success to save source info: {}", request);
         return id;
@@ -111,8 +116,8 @@ public class StreamSourceServiceImpl implements StreamSourceService {
             LOGGER.error("source not found by id={}", id);
             throw new BusinessException(ErrorCodeEnum.SOURCE_INFO_NOT_FOUND);
         }
-        StreamSourceOperator operation = operationFactory.getInstance(SourceType.forType(entity.getSourceType()));
-        StreamSource streamSource = operation.getFromEntity(entity);
+        StreamSourceOperator sourceOperator = operatorFactory.getInstance(SourceType.forType(entity.getSourceType()));
+        StreamSource streamSource = sourceOperator.getFromEntity(entity);
         LOGGER.debug("success to get source by id={}", id);
         return streamSource;
     }
@@ -139,6 +144,30 @@ public class StreamSourceServiceImpl implements StreamSourceService {
     }
 
     @Override
+    public Map<String, List<StreamSource>> getSourcesMap(InlongGroupInfo groupInfo,
+            List<InlongStreamInfo> streamInfos) {
+        String groupId = groupInfo.getInlongGroupId();
+        LOGGER.debug("begin to get source map for groupId={}", groupId);
+        Map<String, List<StreamSource>> result;
+
+        // if the group mode is LIGHTWEIGHT, just get all related stream sources
+        List<StreamSource> streamSources = this.listSource(groupInfo.getInlongGroupId(), null);
+        if (InlongConstants.LIGHTWEIGHT_MODE.equals(groupInfo.getLightweight())) {
+            result = streamSources.stream()
+                    .collect(Collectors.groupingBy(StreamSource::getInlongStreamId, HashMap::new,
+                            Collectors.toCollection(ArrayList::new)));
+        } else {
+            // if the group mode is NORMAL, needs to get the cached MQ sources
+            String sourceType = groupInfo.getMqType();
+            StreamSourceOperator sourceOperator = operatorFactory.getInstance(SourceType.forType(sourceType));
+            result = sourceOperator.getSourcesMap(groupInfo, streamInfos, streamSources);
+        }
+
+        LOGGER.debug("success to get source map, size={}, groupInfo={}", result.size(), groupInfo);
+        return result;
+    }
+
+    @Override
     public PageInfo<? extends StreamSource> listByCondition(SourcePageRequest request) {
         Preconditions.checkNotNull(request.getInlongGroupId(), ErrorCodeEnum.GROUP_ID_IS_EMPTY.getMessage());
 
@@ -154,8 +183,8 @@ public class StreamSourceServiceImpl implements StreamSourceService {
         List<StreamSource> responseList = Lists.newArrayList();
         for (Map.Entry<SourceType, Page<StreamSourceEntity>> entry : sourceMap.entrySet()) {
             SourceType sourceType = entry.getKey();
-            StreamSourceOperator operation = operationFactory.getInstance(sourceType);
-            PageInfo<? extends StreamSource> pageInfo = operation.getPageInfo(entry.getValue());
+            StreamSourceOperator sourceOperator = operatorFactory.getInstance(sourceType);
+            PageInfo<? extends StreamSource> pageInfo = sourceOperator.getPageInfo(entry.getValue());
             if (null != pageInfo && CollectionUtils.isNotEmpty(pageInfo.getList())) {
                 responseList.addAll(pageInfo.getList());
             }
@@ -179,13 +208,13 @@ public class StreamSourceServiceImpl implements StreamSourceService {
         InlongGroupEntity groupEntity = groupCheckService.checkGroupStatus(groupId, operator);
 
         String sourceType = request.getSourceType();
-        StreamSourceOperator operation = operationFactory.getInstance(SourceType.forType(sourceType));
+        StreamSourceOperator sourceOperator = operatorFactory.getInstance(SourceType.forType(sourceType));
         // Remove id in sourceField when save
         List<StreamField> streamFields = request.getFieldList();
         if (CollectionUtils.isNotEmpty(streamFields)) {
             streamFields.forEach(streamField -> streamField.setId(null));
         }
-        operation.updateOpt(request, groupEntity.getStatus(), operator);
+        sourceOperator.updateOpt(request, groupEntity.getStatus(), operator);
 
         LOGGER.info("success to update source info: {}", request);
         return true;
@@ -244,10 +273,10 @@ public class StreamSourceServiceImpl implements StreamSourceService {
         Preconditions.checkNotNull(entity, ErrorCodeEnum.SOURCE_INFO_NOT_FOUND.getMessage());
         groupCheckService.checkGroupStatus(entity.getInlongGroupId(), operator);
 
-        StreamSourceOperator operation = operationFactory.getInstance(SourceType.forType(entity.getSourceType()));
+        StreamSourceOperator sourceOperator = operatorFactory.getInstance(SourceType.forType(entity.getSourceType()));
         SourceRequest sourceRequest = new SourceRequest();
         CommonBeanUtils.copyProperties(entity, sourceRequest, true);
-        operation.restartOpt(sourceRequest, operator);
+        sourceOperator.restartOpt(sourceRequest, operator);
 
         LOGGER.info("success to restart source info: {}", entity);
         return true;
@@ -262,10 +291,10 @@ public class StreamSourceServiceImpl implements StreamSourceService {
         Preconditions.checkNotNull(entity, ErrorCodeEnum.SOURCE_INFO_NOT_FOUND.getMessage());
         groupCheckService.checkGroupStatus(entity.getInlongGroupId(), operator);
 
-        StreamSourceOperator operation = operationFactory.getInstance(SourceType.forType(entity.getSourceType()));
+        StreamSourceOperator sourceOperator = operatorFactory.getInstance(SourceType.forType(entity.getSourceType()));
         SourceRequest sourceRequest = new SourceRequest();
         CommonBeanUtils.copyProperties(entity, sourceRequest, true);
-        operation.stopOpt(sourceRequest, operator);
+        sourceOperator.stopOpt(sourceRequest, operator);
 
         LOGGER.info("success to stop source info: {}", entity);
         return true;
